% Generated by roxygen2 (4.0.1): do not edit by hand
\name{medalplot}
\alias{medalplot}
\title{Medal plot}
\usage{
medalplot(Q, Qo, A, subset = 1L:nrow(A), Sigma_part = NULL, xsamp = NULL)
}
\arguments{
\item{Q}{sparse or dense precision matrix of the Gaussian random field \code{X}}

\item{Qo}{sparse or dense precision of the observation error \code{e}. Typically this is diagonal}

\item{A}{sparse or dense matrix mapping the observations \code{Z} to the underlying field \code{X} under the model \code{Z = AX + e}}

\item{subset}{a vector identifying which of the rows in \code{A} to use in constructing the medals}

\item{Sigma_part}{(optional) the partial or full posterior covariance matrix \code{Var(X|Z)} (if available)}

\item{xsamp}{(optional)  an \code{n x N} matrix containing the samples of the posterior distribution of \code{X}. \code{N} is the number of samples.}
}
\value{
A data frame with five fields, \code{r1}: the radius of the outer disk, \code{r2}: the radius of the middle disk, \code{r3}: the radius of the inner disk, \code{col_outer}: the colour of the outer disk and \code{col_inner}: the colour of the inner disk.
}
\description{
Computes the radii of the inner, middle and outer disks in a `medal plot'.
}
\details{
The medal plot is a visualisation tool for large-scale latent Gaussian models of the form \code{Z = AX + e}
where \code{Z} are the observations, \code{A} is a mapping matrix, \code{X} is the collection of hidden states and
\code{e} is the additive noise with variance \code{T}. The medal plot reveals the relation of the uncertainty of the update on linear combinations
of \code{Y = LX} (where \code{Y} is a subet of the rows of \code{A}) with respect to the observation uncertainty \code{T} and the prior uncertainty \code{Var(X)}.
The medal also highlights the effect of the field's dependence structure on uncertainty reduction. The plot consists of a set of concentric disks with the following properties:
\itemize{
   \item The radius of the outer disk is proportional to either the square root of the observation variance \code{T} or the prior variance
   \code{Var(X)}, whichever is the smaller (note that the lower of these two is a lower bound for the posterior variance in a pure Gaussian system).
   \item The colour of the outer disk is blue if the disk represents the observation uncertainty, red the prior uncertainty.
   \item The radius of the middle disk (in white), corresponding to \code{Var(Yi | Zi)}. i.e. the uncertainty on \code{Yi} if only \code{Zi} were present (or, alternatively, there was no dependence structure in \code{X}).
   \item The inner disk is proportional to the updated variance \code{Var(Y|Z)} and is gold in colour.
}
}
\examples{
### Define the Matern function
Matern <- function(r=0:100,nu=3/2,var=1,kappa=0.1) {
K <- var/((2^(nu-1))*gamma(nu))*(kappa*abs(r))^nu*besselK(kappa*abs(r),nu=nu)
    diag(K) = var
return(K) }

### Construct a valid covariance matrix on a grid of 20 cells
n <- 20
s <- 1:n
S <- Matern(as.matrix(dist(s)), nu = 3/2,var=4, kappa = 0.1)
Q <- chol2inv(chol(S))

### Observations and mapping matrix
sy <- 1:n
sigmav <- rep(1,n)
Qo <- diag(1/sigmav^2)
ny <- length(sy)
A <- diag(rep(1,n))

### Medal plot function
M <- medalplot(Q=Q,Qo=Qo,A=A)
}
\author{
Andrew Zammit-Mangion and Jonathan C. Rougier
}
\references{
Jonathan C. Rougier, Andrew Zammit-Mangion and Nana Schoen (2014). Visualisation for large-scale Gaussian updates. \url{http://www.maths.bris.ac.uk/~MAZJCR/rougierVLSGU.pdf}
}
\keyword{medal}
\keyword{plot,}
\keyword{uncertainty}
\keyword{visualisation}

